==> 1_Web_Scraping_Tutorial.ipynb <==
## Web Scraping ##

This tutorial shows how to "scrape" data online off of a basic HTML website.

[Just give me the code!](#codecell)

Suppose we wanted to get a list of the top movie quotes from the American Film Institute. You can see the table here: <a>http://www.afi.com/100Years/quotes.aspx</a>. 

We'll use BeautifulSoup as the main workhorse to work around the HTML. 

### HTML Summary ###

Text based websites are usually written using just HTML. It is a front-end language which uses tags to separate different styles. For example, to make something appear in bold, you would write:

<b>BOLD TEXT WOULD BE WRITTEN HERE</b>

Similarly, a link, like the site http://www.afi.com/100Years/quotes.aspx will appear as blue and underlined only if you write it as:

<a> http://www.afi.com/100Years/quotes.aspx</a>

Tables are typically written as the following:
<TABLE>
  <tr>
  <td> First Row, First Column Text <\td>
  <td> First Row, Second Column Text <\td>
  <td> First Row, Third Column Text <\td>
  <tr>
  <tr>
  <td> Second Row, First Column Text <\td>
  <td> Second Row, Second Column Text <\td>
  <td> Second Row, Third Column Text <\td>
  <tr>
<\TABLE>

You can probably guess that <tr> means a row is about to start, and </tr> indicates the end of the row.
Likewise, <td> indicates the beginning of an element of the row, and </td> indicates the end of that element.

### Go ahead, make my day. ###

So now, let's get the HTML of the AFI website above.

We'll start by importing the Beautiful Soup module to parse the HTML, urllib2 to connect to the internet, and pandas to save our data neatly.

from bs4 import BeautifulSoup
import urllib2
import pandas as pd

def get_HTML(url):
	req = urllib2.Request(url, headers={ 'User-Agent': 'Mozilla/5.0' })
	html = urllib2.urlopen(req).read()
	soup = BeautifulSoup(html, 'html.parser')
	return soup

"""
If you are using Python 3, do the following

1) "import urllib" instead of "import urllib2"
2) To open a webpage, 

def get_HTML(url):
	req = urllib.request.urlopen(url)
	html = req.read() 
	soup = BeautifulSoup(html, 'html.parser')
	return soup
"""

afi_soup = get_HTML("http://www.afi.com/100Years/quotes.aspx") 

<b>get_HTML</b> is a function which will make a request to our url (<b>req</b>), retrieve the html by reading all the text (<b>html</b>, and make it a BeautifulSoup instance to make it easy to parse the HTML for what we want.

<b>afi_soup</b> is a bunch of html text. 

A nice way to find our target HTML tags (that is, either "b" for bold text, "a" for website links, "table" for tables, etc.) is to go to the website on Chrome or Firefox, right-click and view the page source. Then CTRL+F to search for the table, (like search for "Frankly, my dear, I don't give a damn"). You will see the following:

OTHER STUFF...


<TABLE width="530" border=1 cellspacing=0 cellpadding=2 style="border-collapse:collapse;border:solid black .5pt;">
                        <tr style='height:13.0pt'>
  <td   valign=top class="lttext" >
  <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><b>#</b></span></p>
  </td>
    <td   valign=top  class="lttext">
  <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><b>QUOTE</b></span></p>
  </td>
  <td   valign=top  class="lttext">
 <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><b>MOVIE</b></span></p>
  </td>
    <td   valign=top  class="lttext">
 <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><b>YEAR</b></span></p>
  </td>
 </tr>
  <tr style='height:13.0pt'>
  <td   valign=top sclass="lttext" >
  <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><b>1</b></span></p>
  </td>
  <td   valign=top class="lttext" >
  <p><span style='font-size:9.0pt;font-family:Verdana; color:black'>Frankly, my dear, I don't give a damn.</span></p>
  </td>
  <td   valign=top class="lttext" >
 <p><span style='font-size:9.0pt;font-family:Verdana; color:black'><a href="http://www.afi.com/members/catalog/DetailView.aspx?s=&Movie=1181">GONE WITH THE WIND</a></span></p>
  </td>

So we see the quote, which corresponds to the table, is wrapped in a paragraph tag (this is the "p" tag), which is within an element tag ("td"), which is within a row tag ("tr") which is in the "TABLE". 

So, let's use BeautifulSoup to retrieve only the table elements.

tables = afi_soup.find_all("table")

print(len(tables))

#### Oh no, many tables ####
Within our HTML page, we found 3 tables. So now we need to see if we should only analyze the first, second or third. We can do this by finding the first row of each table, and printing out the first 100 characters of the text.

print("Table 0")
print(tables[0].find("tr").text.strip()[0:200])
print("-----------------------------")
print("Table 1")
print(tables[1].find("tr").text.strip()[0:200])
print("-----------------------------")
print("Table 2")
print(tables[2].find("tr").text.strip()[0:200])
print("-----------------------------")

So we can see that we need to use the 2nd table, <b>tables[1]</b>

quote_table = tables[1]

rows = quote_table.find_all("tr")

sample_row = rows[3]

Let's inspect some characteristics of a sample row, <b>sample_row</b>

print(sample_row)

We see that it still contains a lot of tags. The meat of each row, as mentioned above, is in the "td" tag. So let's extract all the td tags of this row.

print(sample_row.find_all("td"))

We only want the text though. For example, within the row, and for one element we can print out the text by using the <i>text</i> method provided by BeautifulSoup:

print(sample_row.find_all("td")[0].text)
print(sample_row.find_all("td")[1].text)
print(sample_row.find_all("td")[2].text)
print(sample_row.find_all("td")[3].text)

Perfect! Looks like almost everything we want. It would be nice though to put this in a list of 4 elements: the quote number , the quote, the movie, and the year. I'll use the <b>map</b> function from Python to keep only the text of each "td" tag in the row.

row_elements = sample_row.find_all("td")
print(map(lambda x: x.text.strip(),row_elements))

We are now ready to cycle through all the rows, and then loop through all the elements of each row to extract only the quote number, quote, movie and year. I'll use <b>pandas</b> to store them neatly in a dataframe. Note, the first row of the table is just the column headers, which can be fed perfectly into the columns argument of the pandas dataframes.

first = True
for row in quote_table.find_all("tr"):
	row_elements = row.find_all('td')
	row_txt = map(lambda x: x.text.strip(),row_elements)
	if first:
		final_df = pd.DataFrame(columns=row_txt)
		first = False
		continue
	final_df.loc[len(final_df)+1] = row_txt

Let's check out our database:

print(final_df.head())

### More Dynamic Web-Scraping ###

Suppose we also wanted to get more information of each quote, but this required going to the actual link of the movie. For example, we want to see who directed Gone With the Wind (who is Victor Fleming, Alex) which we can see by clicking on the link to the movie. <a> http://www.afi.com/members/catalog/DetailView.aspx?s=&Movie=1181</a>. 

Note that this link shows up in the HTML of the row where Gone with the Wind is located

print(rows[1])

print(rows[1].find_all("td")[2])

We can access specific attributes within an HTML tag by calling it as if it were a dictionary.

rows[1].find_all("td")[2].find("a")["href"]

Thus, to find the link we just need to find the "a" element of the cell and call "href". 

Next, looking at the Gone with the Wind page source, the word <b>Director:</b> is in bold, followed by Victor Flemming. Fortunately, we can use BeautifulSoup to find the tag with <b>Director:</b> and then retrieve the next element by simply using the <b>find_next()</b> method of BeautifulSoup. 

We will do this in two steps.
    1. Run <b>get_HTML</b> on the link to the movie which we retrieve with .find("a")["href"]
    2. Find <b>Director:</b>, then get the next element which will presumably be the name.
    
### NOTE:  I will only do this for the first 10 movies on the link. Data scraping can cause a lot of traffic to the site and they can tell you are a web-scraper by making hundreds of requests rapidly from one IP. 

final_df['Director'] = ""

max_iter = 10
index = 0
for row in quote_table.find_all("tr"):
	if index>max_iter:
		break
	if row.find("a"):
		movie_soup = get_HTML(row.find("a")['href'])
		director = movie_soup.find(text="Director:").find_next().text
		final_df.loc[index,"Director"] = director
	index+=1

print(final_df.head())

Lastly, we can save our database nicely with Pandas as a Stata, Excel, CSV or R file for text analysis

final_df.to_csv("movie_quotes.csv",index=False, encoding='utf-8')

If you want more information, check out the BeautifulSoup docs <a>https://www.crummy.com/software/BeautifulSoup/bs4/doc</a>.

If you need to datascrape java websites or non-html sites, I recommend Selenium <a>selenium-python.readthedocs.io/</a>.

<a id='codecell'></a>

#################################
### Author: Paul Soto 		  ###
### 		paul.soto@upf.edu ###
#								#
# This file shows the basic of ##
# BeautifulSoup to datascrape ###
# an HTML based website and save#
# the data as a csv. 			#
#################################

from bs4 import BeautifulSoup
import urllib2
import pandas as pd

def get_HTML(url):
	"""
	This file creates a HTML soup from a given url
	"""
	req = urllib2.Request(url, headers={ 'User-Agent': 'Mozilla/5.0' })
	html = urllib2.urlopen(req).read()
	soup = BeautifulSoup(html, 'html.parser')
	return soup


# Get HTML Soup
afi_soup = get_HTML("http://www.afi.com/100Years/quotes.aspx")

# Isolate the table with the  
tables = afi_soup.find_all("table")
quote_table = tables[1]

# Loop through each row, retrieving the 
# quote ID number, quote, movie and year
first = True
for row in quote_table.find_all("tr"):
	# Get row elements
	row_elements = row.find_all('td')
	# Strip unnecessary text
	row_txt = map(lambda x: x.text.strip(),row_elements)
	if first:
		final_df = pd.DataFrame(columns=row_txt)
		first = False
		continue
	# Add to dataset
	final_df.loc[len(final_df)+1] = row_txt

# Clicking movie link to get the director
final_df['Director'] = ""
max_iter = 10
index = 0
for row in quote_table.find_all("tr"):
	if index>max_iter:
		break
	if row.find("a"):
		movie_soup = get_HTML(row.find("a")['href'])
		director = movie_soup.find(text="Director:").find_next().text
		final_df.loc[index,"Director"] = director
	index+=1

# Export to CSV
final_df.to_csv("movie_quotes.csv",index=False)
