==> 3_EM_Algorithm_Tutorial.ipynb <==
# Expectation Maximization (EM) Algorithm Applied to Text

This tutorial summarizes how the EM algorithm can be applied to texts. It is based on Stephen Hansen's lecture notes <a>https://sekhansen.github.io/pdf_files/lecture2.pdf</a> and text-mining module <a>https://github.com/sekhansen/text-mining-tutorial<a/>.

[**Just Give Me the Code**](#gimmethecode)

First, suppose we have 3 documents:

data = [["rugby","football","competition","ball","games"],
        ["macro","economics","io","competition","games","econometrics"],
        ["business","economics","stocks","bonds","NYSE"]]

For example, document 1 is a <i>sports</i> document, document 2 is related to <i>economics</i>, and document 3 is related to <i>business</i>. 

**Goal**
1. Obtain Topic Distribution <br>
The goal is to find the distribution of K topics. That is, in our corpus what is the probability a random document is related to topic 1 or topic 2 or ... topic K? This is represented by $\rho$=$(\rho_1,\rho_2,...\rho_K)$ where $1\geq\rho_k\geq0$. This is just a vector of size K. 
<br>
<br>
2. Obtain Word Distrubition <br>
Given a topic $k$, what is the probability that the document contains any word in our vocabulary of size V. This is represented by $\mathbf\beta_k$=($\beta_{1,k},\beta_{2,k},...,\beta_{V,k})$ where $1\geq\beta_{v,k}\geq0$. We will have K such vectors, so $\mathbf\beta$ will be a matrix of size K by V.
<br>
<br>

We will assume that the probability of observing a document $w_d$ is: $Pr(w_d | \rho, \mathbf\beta)=\Sigma_{k=1}^K(\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}})$, where $x_{v,d}$ is the occurences of word v in document d.


So the probability of observing document 1 is just: <br>$Pr$(["rugby","football","competition","ball","games"] | $\rho, \mathbf\beta)=\rho_{1}(\beta_{ball,1}^{1}*\beta_{bonds,1}^{0}*...)+\rho_{2}(\beta_{ball,2}^{1}*\beta_{bonds,2}^{0}*...)$

We can represent the data in a frequency matrix where each row is a document, and each column is a word the vocabulary. The values correspond to the frequencies a word shows up in a document

import pandas as pd
pd.set_option('display.expand_frame_repr', False)

feature_counts = pd.DataFrame(data=
                [[ 1.,  0.,  0.,  1.,  0.,  1.,  1.,  0.,  1.,  0.],                            
                 [ 0.,  0.,  0.,  1.,  1.,  0.,  1.,  1.,  0.,  0.],
                 [ 0.,  1.,  1.,  0.,  1.,  0.,  0.,  0.,  0.,  1.]], 
                columns=[u'ball', u'bonds', u'business', u'competition', u'economics',                   
                         u'football', u'games', u'macro', u'rugby', u'stocks'])

print(feature_counts)

**Implementation**
The EM algorithm starts by guessing the probability that a document comes from topic $k$. Let's define the number of topics, K, and our initial guess of the distribution of each topic.

We import numpy as this module gives us a lot of useful functions like linear algebra operations and sampling from different types of distributions. 

<b>k</b> is the number of (unknown or latent) topics.

<b>rhos</b> ($\rho$) is the initial "guess" of the distribution of topics. The uniform distribution is a good starting point. We'll guess the probability of a document belonging to topic 1 is 50% and topic 2 is 50%. 

<b>betas</b> ($\mathbf\beta$) we will also guess , the probability matrix where word $v$ appears in topic $k$

import numpy as np

K = 2

D = feature_counts.shape[0]

V = feature_counts.shape[1]

rhos = np.full(K,1/float(K))

betas = np.random.dirichlet(V*[1], K)

##Summary of Algorithm##
$Pr(w_d | \rho, \mathbf\beta)$ is the likelihood of observing an entire document. Though consider that we are summing over topics. So we can separate each $(\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}})$ and interpret this loosely as how much that document <i>likes</i> topic k. If $(\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}})$ is really high, then according to the distributional parameters $\beta$ and $\rho$, document $w_d$ is a good candidate to be in topic $k$. So we'll proceed first by computing a matrix showing how much each document <i>likes</i> each of the $k$ topics.

**Step 1: Expectation Step**
<br> $\text{Document Topic Matrix} 
\begin{equation*}
 =  \begin{vmatrix}
\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc1}} & \rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc1}} \\
\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc2}} & \rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc2}} \\
\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc3}} & \rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc3}} 
\end{vmatrix}
\end{equation*}$

We will normalize this Document Topic Matrix so that the rows sum to 1. Then, each row element $i,j$ can be viewed as a probability that document $i$ belongs to topic $j$. We will call this the weighting matrix, $z$.

Importantly, at the E-Step we should compute the total likelihood to see how likely the entire corpus of documents would occur under the current estimates of $\beta$ and $\rho$. If our likelihood only increases by a small amount, we can be happy assuming the current estimates of $\beta$ and $\rho$. 

**Step 2: Maximization Step**
<br>

$ z \begin{equation*}
 =  \begin{vmatrix}
\frac{\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc1}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc1}}}
& \frac{\rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc1}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc1}}}
 \\
\frac{\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc2}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc2}}}
& \frac{\rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc2}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc2}}} \\
\frac{\rho_{topic1} \Pi_{v=1}^V\beta_{v,topic1}^{x_{v,doc3}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc3}}}
& \frac{\rho_{topic2} \Pi_{v=1}^V\beta_{v,topic2}^{x_{v,doc3}}}{\Sigma_k \rho_{k} \Pi_{v=1}^V\beta_{v,k}^{x_{v,doc3}}}
\end{vmatrix}
\end{equation*}$

We can average the $z$ matrix of probabilities <i>across</i> documents to get anew estimate of $\rho$. 

Similarly, the normalized Document Topic Matrix can be used to update $\beta$. We will simply multiply the z matrix by the document term matrix above (with the counts of each word in each document. The z matrix is DxK, and the frequency count matrix is DxV, so we need to transpose the z matrix. Then will will also normalize it so that the rows of the final KxV matrix sum to 1. 

### E-Step###

First let's see how much each document likes each topic. We do this by computing each documents likelihood, given our initial guess of $\rho$ and $\beta$. We will log these likelihoods then exponentiate them later.

Log-Likelihood of document $w_d$:
$log(Pr(w_d | \rho, \mathbf\beta))=\Sigma_{k=1}^K(log(\rho_k)+ \Sigma_{v=1}^Vlog(\beta_{v,k})*{x_{v,d}})$

dt_temp = np.zeros((D,K))
for doc in range(D):
    for topic in range(K):
        dt_temp[doc,topic] = np.log(rhos[topic]) + (feature_counts.values[doc]*np.log(betas[topic])).sum()

We are in the log world so we need to exponentiate the likelihoods to retrieve $\Sigma_{k=1}(\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}})$

print(np.exp(dt_temp))

Note the extremely small numbers. We will eventually need to calculate the complete likelihood over all documents, that is $l(X|\rho\beta) = \Sigma_{d=1}^{D} log(\Sigma_{k=1}^K\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}})=\Sigma_{d=1}^{D} log(\Sigma_{k=1}^Ke^{(log(\rho_k)+ \Sigma_{v=1}^Vlog(\beta_{v,k})*{x_{v,d}})})$. Taking the logarithms of such small numbers, as in the print above, will lead to <b>numerical underflow</b>. Fortunately, we can use the <b>log-sum trick</b>, which is factoring out the largest exponent:

$log\Sigma_{c}e^{b_c} = log[(\Sigma_{c}e^{b_c-B})e^B]=[log(\Sigma_ce^{b_c-B})]+B$, where $B$ is the largest $b_c$

def complete_ll(dt_matrix):
    """
    Computes the total likelihood using
    Pr(w_d) = sigma_k rho_k * Pi_v (beta_k,v)

    and ll = 
    """
    ### Use log-sum trick
    dt_tempZ = dt_matrix - np.max(dt_matrix,1)[:,np.newaxis]
    ll = np.log(np.exp(dt_tempZ).sum(1)).sum()+dt_temp.max(1).sum()
    return ll

print(complete_ll(dt_temp))


### M-Step###

Now we need to update our initial guesses of $\rho$ and $\beta$. 

1. First, let's update $\rho$. 

We will use the formula:

$\rho_k^{new}=\frac{\Sigma_d\hat{z_{d,k}}}{\Sigma_k\Sigma_d\hat{z_{d,k}}}$ where $\hat{z_{d,k}}=\frac{\rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}}}{\Sigma_d \rho_k \Pi_{v=1}^V\beta_{v,k}^{x_{v,d}}}$

probs = np.exp(dt_temp)/np.exp(dt_temp).sum(1)[:,np.newaxis]

print(probs)

Each row shows the distribution of a document across the K topics, where the rows sum to 1. For example, the top left element is how much ["rugby","football","competition","ball","games"] likes topic 1. 

We can just average each column across rows to come up with the new values of $\rho$.

rhos = probs.sum(0)/D

print(rhos)

2. Second, let's update $\beta$

We will use the formula 
$\beta_{k,v}^{new}=\frac{\Sigma_d\hat{z_{d,k}}*x_{d,v}}{\Sigma_d\hat{z_{d,k}}*\Sigma_vx_{d,v}}$



betas_temp = np.dot(probs.T,feature_counts.values)
betas=betas_temp/betas_temp.sum(1)[:,np.newaxis]

print(betas)

And that's one iteration of the algorithm. The E-step and the M-step would be repeated until the log-likehood increases very litle. We can implement the entire code in the form of a class as follows:

<a id='gimmethecode'></a>

#################################
### Author: Paul Soto 		  ###
### 		paul.soto@upf.edu ###
#								#
# This file is a class to #######
# perform the EM algorithm ######
# on a multinomial distribution##
#################################

import numpy as np


class EM_algo():
	""""
	Algorithm for the multinomial mixture model
	"""
	def __init__(self, data, K=2,
					prior_rhos=None, 
					prior_betas=None):
		self.D = data.shape[0]
		self.V = data.shape[1]
		self.K = K
		self.data = data
		if prior_rhos==None:
			self.rhos = np.full(self.K, 1/float(self.K))
		else:
			self.rhos = prior_rhos
		if prior_betas==None:
			self.betas = np.random.dirichlet(self.V*[1], self.K)
		else:
			self.betas =prior_betas 
		self.ll_history = []

	def E_step(self):

		def complete_ll(dt_matrix):
			"""
			Computes the total likelihood using
			Pr(w_d) = sigma_k rho_k * Pi_v (beta_k,v)

			and ll = 
			"""
			### Use log-sum trick
			dt_tempZ = dt_matrix - np.max(dt_matrix,1)[:,np.newaxis]
			ll = np.log(np.exp(dt_tempZ).sum(1)).sum()+dt_temp.max(1).sum()
			return ll

		### Get likelihood for each document-topic
		dt_temp = np.zeros((self.D,self.K))
		for doc in range(self.D):
			for topic in range(self.K):
				dt_temp[doc,topic] = np.log(self.rhos[topic]) + \
									(self.data[doc]*np.log(self.betas[topic])).sum()
		### Compute complete log-likelihood
		self.ll = complete_ll(dt_temp)
		self.ll_history.append(self.ll)

		### normalize the Zs
		self.zs = np.exp(dt_temp)/np.exp(dt_temp).sum(1)[:,np.newaxis]


	def M_step(self):
		### Update rhos
		self.rhos = self.zs.sum(0)/self.D
		### Update betas
		betas_temp = np.dot(self.zs.T,self.data)
		self.betas=betas_temp/betas_temp.sum(1)[:,np.newaxis]

	def iterate(self,tolerance=0.001,max_iter = 500):
		count = 0
		while True:
			self.E_step()
			self.M_step()
			if len(self.ll_history)==1:
				continue
			if (self.ll-self.ll_history[-2])<tolerance:
				print("Arrived at target tolerance!")
				break
			if (count>max_iter):
				print("Exceeded max_iterations!")
				break
			count+=1

em_instance = EM_algo(feature_counts.values,K=2)

em_instance.iterate()

print(em_instance.rhos)
print(em_instance.betas)
